Fix unit testing, docs are still broke
