+++ gwdetchar/nagios/__init__.py
@@ -1,0 +1,24 @@
+# coding=utf-8
+# Copyright (C) Alex Urban (2019)
+#
+# This file is part of the GW DetChar python package.
+#
+# GW DetChar is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# GW DetChar is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with GW DetChar.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Utilities for providing status updates to Nagios
+"""
+
+from .core import *
+
+__author__ = 'Alex Urban <alexander.urban@ligo.org>'
